Feature/update user

+++ server/controllers/books.js
@@ -1,5 +1,4 @@
 import { Book, BorrowedBook, BookCategory } from '../models';
-import notifyAdmin from '../helpers/notifyAdmin';
 
 /**
  * Fetch all books that match a catagory from database
@@ -13,8 +12,7 @@ const filterBooksByCategory = (req, res) => {
   const category = req.query.category;
   BookCategory.findAll({ where: { category } })
     .then((books) => {
-      const message = books.length ? '' :
-        'No books in match the requested category';
+      const message = books.length ? '' : 'No books in match the requested category';
       res.status(200).send({
         success: true,
         data: books,
@@ -38,17 +36,22 @@ export default {
    * @return {undefined}
    */
   addCategory(req, res) {
-    return BookCategory
-      .create(req.body)
-      .then(category => res.status(201).send({
-        success: true,
-        message: `Successfully added new category, ${category.category},
-          to Library`,
-      }))
-      .catch(error => res.status(500).send({
-        success: false,
-        error
-      }));
+    if (req.user && req.user.isAdmin) {
+      return BookCategory
+        .create(req.body)
+        .then(category => res.status(201).send({
+          success: true,
+          message: `Successfully added new category, ${category.category}, to Library`,
+        }))
+        .catch(error => res.status(500).send({
+          success: false,
+          error
+        }));
+    }
+    res.status(401).send({
+      success: false,
+      message: 'Unauthorized access',
+    });
   },
   /**
    * Fetch Bppk Categories.
@@ -81,17 +84,23 @@ export default {
    * @return {undefined}
    */
   createBook(req, res) {
-    return Book
-      .create(req.body)
-      .then(book => res.status(201).send({
-        success: true,
-        message: `Successfully added ${book.title} to Library`,
-        data: book,
-      }))
-      .catch(error => res.status(500).send({
-        success: false,
-        error
-      }));
+    if (req.user && req.user.isAdmin) {
+      return Book
+        .create(req.body)
+        .then(book => res.status(201).send({
+          success: true,
+          message: `Successfully added ${book.title} to Library`,
+          data: book,
+        }))
+        .catch(error => res.status(500).send({
+          success: false,
+          error
+        }));
+    }
+    res.status(401).send({
+      success: false,
+      message: 'Unauthorized access',
+    });
   },
 
   /**
@@ -167,21 +176,28 @@ export default {
    */
   editBookInfo(req, res) {
     const id = req.params.id;
-    Book.update(
-      req.body,
-      { where: { id },
-        returning: true,
-        plain: true,
-      })
-      .then(book => res.status(200).send({
-        success: true,
-        book: book[1],
-        message: `${book[1].title} was successfully updated`
-      }))
-      .catch(error => res.status(500).send({
+    if (req.user && req.user.isAdmin) {
+      Book.update(
+        req.body,
+        { where: { id },
+          returning: true,
+          plain: true,
+        })
+        .then(book => res.status(200).send({
+          success: true,
+          book: book[1],
+          message: `${book[1].title} was successfully updated`
+        }))
+        .catch(error => res.status(500).send({
+          success: false,
+          error,
+        }));
+    } else {
+      res.status(401).send({
         success: false,
-        error,
-      }));
+        message: 'Unauthorized access',
+      });
+    }
   },
 
   /**
@@ -194,15 +210,22 @@ export default {
    */
   deleteBook(req, res) {
     const id = req.params.id;
-    Book.destroy({ where: { id } })
-      .then(() => res.status(200).send({
-        success: true,
-        message: 'Successfully deleted book from database',
-      }))
-      .catch(error => res.status(500).send({
+    if (req.user && req.user.isAdmin) {
+      Book.destroy({ where: { id } })
+        .then(() => res.status(200).send({
+          success: true,
+          message: 'Successfully deleted book from database',
+        }))
+        .catch(error => res.status(500).send({
+          success: false,
+          error,
+        }));
+    } else {
+      res.status(401).send({
         success: false,
-        error,
-      }));
+        message: 'Unauthorized access',
+      });
+    }
   },
 
   /**
@@ -237,8 +260,7 @@ export default {
             if (borrowed && borrowed.returned === false) {
               res.status(403).send({
                 success: false,
-                message: `You currently have this book.
-                  Return it before trying to borrow it again`,
+                message: 'You currently have this book. Return it before trying to borrow it again',
               });
               return;
             } else if (borrowed && borrowed.returned === true) {
@@ -246,11 +268,9 @@ export default {
               borrowed.save();
               book.total -= 1;
               book.save();
-              notifyAdmin('borrow', book.title, req.user.username);
               res.status(200).send({
                 success: true,
-                message: `You have successfully borrowed ${book.title} again.
-                  Check your dashboard to read it`,
+                message: `You have successfully borrowed ${book.title} again. Check your dashboard to read it`,
               });
               return;
             }
@@ -261,14 +281,10 @@ export default {
                 book.total -= 1;
                 book.save();
               })
-              .then(() => {
-                notifyAdmin('borrow', book.title, req.user.username);
-                res.status(200).send({
-                  success: true,
-                  message: `You have successfully borrowed ${book.title}.
-                    Check your dashboard to read it`,
-                });
-              })
+              .then(() => res.status(200).send({
+                success: true,
+                message: `You have successfully borrowed ${book.title}. Check your dashboard to read it`,
+              }))
               .catch((error) => {
                 res.status(500).send({
                   success: false,
@@ -309,7 +325,6 @@ export default {
                 return book;
               })
               .then((book) => {
-                notifyAdmin('return', book.title, req.user.username);
                 res.status(201).send({
                   success: true,
                   message: `You have successfully returned ${book.title}`,
@@ -320,8 +335,7 @@ export default {
         }
         res.status(400).send({
           success: false,
-          message: `This book is currently not on your list.
-            You have either returned it or never borrowed it`
+          message: 'This book is currently not on your list. You have either returned it or never borrowed it'
         });
       })
       .catch(error => res.status(500).send({
